requirements.txt
attrs==23.1.0
ax-platform==0.2.4
backcall==0.2.0
bleach==6.0.0
botorch==0.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
cmdstanpy==0.9.5
convertdate==2.4.0
cycler==0.11.0
Cython==3.0.2
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.3.0
ephem==4.1.4
exceptiongroup==1.1.3
fbprophet==0.7.1
fonttools==4.38.0
gpytorch==1.8.1
hijri-converter==2.3.1
holidays==0.24
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==7.34.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
kaggle==1.5.16
kats==0.2.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
llvmlite==0.39.1
loguru==0.7.2
LunarCalendar==0.0.9
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
multipledispatch==1.0.0
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.3.5
parameterized==0.9.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
plotly==5.16.1
pluggy==1.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
Pygments==2.16.1
pymannkendall==1.4.3
PyMeeus==0.5.12
pyparsing==3.1.1
pystan==2.19.1.1
pytest==7.4.2
pytest-mpl==0.16.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
setuptools-git==1.2
six==1.16.0
sklearn==0.0.post9
statsmodels==0.12.2
tenacity==8.2.3
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
tqdm==4.66.1
traitlets==5.9.0
typeguard==4.1.2
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
zipp==3.15.0
